+++ apps/read-values.yaml
@@ -1,0 +1,33 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: read-values
+spec:
+  selector:
+    matchLabels:
+      app: read-values
+  template:
+    metadata:
+      labels:
+        app: read-values
+    spec:
+      containers:
+      - name: read-values
+        image: matheuscruzdev/read-values
+        resources:
+          limits:
+            memory: "128Mi"
+            cpu: "500m"
+        ports:
+        - containerPort: 8080
+---
+apiVersion: v1
+kind: Service
+metadata:
+  name: read-values-svc
+spec:
+  selector:
+    app: read-values
+  ports:
+  - port: 8080
+    targetPort: 8080
