+++ dapr/subscription.yaml
@@ -1,0 +1,8 @@
+apiVersion: dapr.io/v1alpha1
+kind: Subscription
+metadata:
+  name: notifications-subscription
+spec:
+  topic: notifications 
+  route: /notifications
+  pubsubname: notifications-pubsub
