--- pubsub.yaml
@@ -1,21 +1,0 @@
-apiVersion: dapr.io/v1alpha1
-kind: Component
-metadata:
-  name: notifications-pubsub
-spec:
-  type: pubsub.redis
-  version: v1
-  metadata:
-  # These settings will work out of the box if you use `helm install
-  # bitnami/redis`.  If you have your own setup, replace
-  # `redis-master:6379` with your own Redis master address, and the
-  # Redis password with your own Secret's name. For more information,
-  # see https://docs.dapr.io/operations/components/component-secrets .
-  - name: redisHost
-    value: redis-master:6379
-  - name: redisPassword
-    secretKeyRef:
-      name: redis
-      key: redis-password
-auth:
-  secretStore: kubernetes
