--- read-values/handle_test.go
@@ -1,26 +1,0 @@
-package function
-
-import (
-	"context"
-	"net/http"
-	"net/http/httptest"
-	"testing"
-)
-
-// TestHandle ensures that Handle executes without error and returns the
-// HTTP 200 status code indicating no errors.
-func TestHandle(t *testing.T) {
-	var (
-		w   = httptest.NewRecorder()
-		req = httptest.NewRequest("GET", "http://example.com/test", nil)
-		res *http.Response
-	)
-
-	Handle(context.Background(), w, req)
-	res = w.Result()
-	defer res.Body.Close()
-
-	if res.StatusCode != 200 {
-		t.Fatalf("unexpected response code: %v", res.StatusCode)
-	}
-}
