+++ read-values/main.go
@@ -1,0 +1,78 @@
+package main
+
+import (
+	"context"
+	"encoding/json"
+	"net"
+	"net/http"
+	"os"
+	"strconv"
+
+	dapr "github.com/dapr/go-sdk/client"
+	"github.com/go-chi/chi"
+)
+
+var STATE_STORE_NAME = "statestore"
+var daprClient dapr.Client
+var DAPR_HOST = "my-ambient-dapr-ambient.default.svc.cluster.local"
+var DAPR_PORT = "50001"
+
+type MyValues struct {
+	Values []string
+}
+
+func main() {
+	r := chi.NewRouter()
+	r.Get("/", Handle)
+	r.Get("/health/{endpoint:readiness|liveness}", func(w http.ResponseWriter, r *http.Request) {
+		json.NewEncoder(w).Encode(map[string]bool{"ok": true})
+	})
+	port := GetenvOrDefault("APP_PORT", "8080")
+	http.ListenAndServe(":"+port, r)
+}
+
+// Handle an HTTP Request.
+func Handle(res http.ResponseWriter, req *http.Request) {
+
+	ctx := context.Background()
+
+	daprClient, err := dapr.NewClientWithAddress(net.JoinHostPort(DAPR_HOST, DAPR_PORT))
+	if err != nil {
+		panic(err)
+	}
+
+	result, err := daprClient.GetState(ctx, STATE_STORE_NAME, "values", nil)
+	if err != nil {
+		panic(err)
+	}
+	myValues := MyValues{}
+	json.Unmarshal(result.Value, &myValues)
+
+	var total int
+	var count int
+	for _, v := range myValues.Values {
+		intVar, _ := strconv.Atoi(v)
+		total += intVar
+		count++
+	}
+
+	var avg float64
+	avg = float64(total / count)
+	respondWithJSON(res, http.StatusOK, avg)
+}
+
+func respondWithJSON(w http.ResponseWriter, code int, payload interface{}) {
+	response, _ := json.Marshal(payload)
+
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(code)
+	w.Write(response)
+}
+
+func GetenvOrDefault(envName, defaultValue string) string {
+	v := os.Getenv(envName)
+	if v != "" {
+		return v
+	}
+	return defaultValue
+}
