+++ subscriber/go.mod
@@ -2,4 +2,4 @@ module github.com/salaboy/subscriber
 
 go 1.14
 
-require github.com/gorilla/mux v1.8.0
+require github.com/go-chi/chi v1.5.4
