--- subscriber/kservice.yaml
@@ -1,21 +1,0 @@
-apiVersion: serving.knative.dev/v1
-kind: Service
-metadata:
-  name: subscriber
-  namespace: default
-spec:
-  template:
-    spec:
-      containers:
-        - image: ko://github.com/salaboy/subscriber
-          env:
-            - name: TARGET
-              value: "Go Sample v1"
-          livenessProbe:
-            httpGet:
-              path: /health/liveness
-              port: 8080
-          readinessProbe:
-            httpGet:
-              path: /health/readiness
-              port: 8080
