--- subscription.yaml
@@ -1,8 +1,0 @@
-apiVersion: dapr.io/v1alpha1
-kind: Subscription
-metadata:
-  name: notifications-subscritpion
-spec:
-  topic: notifications 
-  route: /notifications
-  pubsubname: notifications-pubsub
