+++ write-values/Dockerfile
@@ -1,0 +1,5 @@
+FROM alpine
+
+COPY main main
+
+ENTRYPOINT [ "./main" ]
