--- write-values/func.yaml
@@ -1,26 +1,0 @@
-specVersion: 0.34.0
-name: write-values
-runtime: go
-registry: docker.io/salaboy
-image: docker.io/salaboy/write-values:latest
-imageDigest: sha256:2534441dd16c97b1d721f1b6b94829ed1e63a0f780711f1141ce7424e5f3bd5d
-created: 2023-02-20T11:13:56.924349Z
-invocation:
-  format: http
-build:
-  buildpacks:
-  - paketo-buildpacks/go-dist
-  - ghcr.io/boson-project/go-function-buildpack:tip
-  builder: pack
-  buildEnvs: []
-run:
-  volumes: []
-  envs: []
-deploy:
-  namespace: default
-  annotations: {}
-  options: {}
-  labels: []
-  healthEndpoints:
-    liveness: /health/liveness
-    readiness: /health/readiness
