+++ write-values/go.mod
@@ -1,3 +1,8 @@
 module function
 
 go 1.14
+
+require (
+	github.com/dapr/go-sdk v1.7.0
+	github.com/go-chi/chi v1.5.4
+)
